+++ CHANGELOG.md
@@ -38,6 +38,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Removed
 
+- Drop the old `Checkbox` model and mapper (#3818)
 - Drop the old `TimeSlot` model and mapper (#3815, #3817)
 - Drop icons for categories (#3777)
 - Drop the unused `SpeakerBagBuilder` (#3780)